fix(metamodel): widen Integer and Long AST fields to f64 (OD-3)
concerto-metamodel's codegen read Integer and Long AST fields as i32/i64,
so models the TS reference loads (a Long bound above i64::MAX, an
Integer domain bound outside i32's range, a float in a location field)
failed to load in Rust. TS reads these fields as plain JS numbers, so
widen the generated types to f64 to match, keeping values exactly as TS
sees them.

Minimal core call-site fixes for the resulting type change:
- check_size/check_length compare against 0.0 instead of an integer
  literal.
- a size-validator test asserts against f64 literals.

Adds tests in concerto-metamodel and concerto-core covering the
previously-rejected cases (an Integer/Long domain bound beyond
i32::MAX/i64::MAX, a fractional domain bound, a float Position field),
checked against the frozen TS 5.0.0 reference (migration/oracle/reference
in the /home/user/concerto workspace: ModelManager.fromAst loading the
same ASTs returns the same values), and records the resulting
serialization-formatting divergence (DV-005) in DIVERGENCES.md.

+++ concerto-metamodel/build.rs
@@ -411,8 +411,13 @@ impl<'a> TypeTable<'a> {
         let mut rust = match kind {
             "StringProperty" => "String".to_string(),
             "BooleanProperty" => "bool".to_string(),
-            "IntegerProperty" => "i32".to_string(),
-            "LongProperty" => "i64".to_string(),
+            // Integer and Long AST fields are widened to `f64`: TS reads
+            // them as plain JS numbers (53-bit semantics), so a Long bound
+            // above `i64::MAX`, an Integer bound that overflows `i32`, or a
+            // float in a location field all load, matching the TS reference
+            // (OD-3). `i32`/`i64` would reject models TS accepts.
+            "IntegerProperty" => "f64".to_string(),
+            "LongProperty" => "f64".to_string(),
             "DoubleProperty" => "f64".to_string(),
             "ObjectProperty" => {
                 let target = owner.resolve(&property["type"]);
